refactor: add helper functions on block schema
