+++ schemas/block.ts
@@ -1,9 +1,11 @@
-export default {
+import { defineType } from "sanity";
+
+export const block = defineType({
+  title: "Contenido",
   name: "content",
   type: "array",
-  title: "Contenido",
   of: [
     { type: "block", title: "Texto", name: "blockscontent" },
     { type: "img", title: "Image", name: "blocksimage" },
   ],
-};
+});
