+++ schemas/index.ts
@@ -1,9 +1,9 @@
 // TODO: Migrate schemas from './studio/schemas' to './schemas'
 import about from "studio/schemas/about";
 import accessibilityDeclaration from "studio/schemas/accessibility-declaration";
-import block from "studio/schemas/block";
 import characteristics from "studio/schemas/characteristics";
 import characteristics_list from "studio/schemas/characteristics_list";
+import collaborate from "studio/schemas/collaborate";
 import document from "studio/schemas/document";
 import documents from "studio/schemas/documents";
 import faq from "studio/schemas/faq";
@@ -13,7 +13,7 @@ import link from "studio/schemas/link";
 import member from "studio/schemas/member";
 import toy from "studio/schemas/toy";
 import video from "studio/schemas/video";
-import collaborate from "studio/schemas/collaborate";
+import { block } from "./block";
 
 const schemas = [
   /* Your types here! */
