+++ studio/schemas/accessibility-declaration.ts
@@ -1,4 +1,4 @@
-import block from "./block";
+import { block } from "../../schemas/block";
 
 export default {
   name: "accessibilityDeclaration",
