+++ studio/schemas/schema.ts
@@ -5,16 +5,16 @@ import createSchema from "part:@sanity/base/schema-creator";
 // Then import schema types from any plugins that might expose them
 import schemaTypes from "all:part:@sanity/base/schema-type";
 
+import { block } from "../../schemas/block";
 import about from "./about";
 import accessibilityDeclaration from "./accessibility-declaration";
-import block from "./block";
 import characteristics from "./characteristics";
 import characteristics_list from "./characteristics_list";
+import collaborate from "./collaborate";
 import document from "./document";
 import documents from "./documents";
 import faq from "./faq";
 import homePage from "./home-page";
-import collaborate from "./collaborate";
 import image from "./image";
 import link from "./link";
 import member from "./member";
